+++ src/frontend/apps/impress/src/features/docs/doc-management/components/index.ts
@@ -1,2 +1,1 @@
 export * from './ModalRemoveDoc';
-export * from './ModalShare';
