+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useDeleteDocAccess.ts
@@ -6,7 +6,7 @@ import {
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
 import { KEY_DOC, KEY_LIST_DOC } from '@/features/docs/doc-management';
-import { KEY_LIST_USER } from '@/features/docs/members/members-add';
+import { KEY_LIST_USER } from '@/features/docs/doc-share';
 import { useBroadcastStore } from '@/stores';
 
 import { KEY_LIST_DOC_ACCESSES } from './useDocAccesses';
