+++ src/frontend/apps/impress/src/features/docs/doc-share/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useTranslatedShareSettings';
+export * from './useWhoAmI';
