+++ src/frontend/apps/impress/src/features/docs/doc-share/index.ts
@@ -1,0 +1,2 @@
+export * from './api';
+export * from './hooks';
