+++ src/frontend/apps/impress/src/features/docs/doc-share/types.tsx
@@ -1,4 +1,21 @@
 import { User } from '@/core/auth';
+import { Role } from '@/features/docs';
+
+export interface Invitation {
+  id: string;
+  role: Role;
+  document: string;
+  created_at: string;
+  is_expired: boolean;
+  issuer: string;
+  email: string;
+  abilities: {
+    destroy: boolean;
+    retrieve: boolean;
+    partial_update: boolean;
+    update: boolean;
+  };
+}
 
 export enum OptionType {
   INVITATION = 'invitation',
