--- src/frontend/apps/impress/src/features/docs/members/invitation-list/api/index.ts
@@ -1,4 +1,0 @@
-export * from './useCreateDocInvitation';
-export * from './useDeleteDocInvitation';
-export * from './useDocInvitations';
-export * from './useUpdateDocInvitation';
