--- src/frontend/apps/impress/src/features/docs/members/invitation-list/components/index.ts
@@ -1,1 +1,0 @@
-export * from './InvitationList';
