--- src/frontend/apps/impress/src/features/docs/members/invitation-list/conf.ts
@@ -1,1 +1,0 @@
-export const PAGE_SIZE = 20;
