--- src/frontend/apps/impress/src/features/docs/members/members-add/api/index.ts
@@ -1,2 +1,0 @@
-export * from './useCreateDocAccess';
-export * from './useUsers';
