--- src/frontend/apps/impress/src/features/docs/members/members-add/components/index.ts
@@ -1,2 +1,0 @@
-export * from './AddMembers';
-export * from './ChooseRole';
