--- src/frontend/apps/impress/src/features/docs/members/members-add/index.ts
@@ -1,2 +1,0 @@
-export * from './api';
-export * from './components';
