--- src/frontend/apps/impress/src/features/docs/members/members-list/api/index.ts
@@ -1,3 +1,0 @@
-export * from './useDeleteDocAccess';
-export * from './useDocAccesses';
-export * from './useUpdateDocAccess';
