--- src/frontend/apps/impress/src/features/docs/members/members-list/components/index.ts
@@ -1,1 +1,0 @@
-export * from './MemberList';
